feat : util 함수로 공통부 추출
