+++ trippy-client/src/stores/airTicketStore.js
@@ -1,5 +1,6 @@
 import { defineStore } from "pinia";
 import { fetchAirTicketAll } from "@/api/airTicket";
+import { extractDateAndDayOfWeek, normalizeTime } from "@/assets/utils";
 
 export const useAirTicketStore = defineStore("airTicket", {
   state: () => ({
@@ -21,40 +22,10 @@ export const useAirTicketStore = defineStore("airTicket", {
       if (ticket) ticket.used = used;
     },
 
-    // 날짜에서 요일 추출 헬퍼 함수
-    extractDateAndDayOfWeek(departureDateString) {
-      // "2025-08-28(일)" 형식에서 날짜와 요일 분리
-      const match = departureDateString.match(/^(\d{4}-\d{2}-\d{2})\((.)\)$/);
-      if (match) {
-        return {
-          date: match[1], // "2025-08-28"
-          dayOfWeek: match[2], // "일"
-        };
-      }
-      // 매칭 실패시 fallback
-      return {
-        date: departureDateString.split("(")[0] || departureDateString,
-        dayOfWeek: "미정",
-      };
-    },
-
-    // 시간 형식 정규화 (0:30 -> 00:30)
-    normalizeTime(timeString) {
-      if (!timeString) return "00:00";
-
-      const parts = timeString.split(":");
-      if (parts.length === 2) {
-        const hours = parts[0].padStart(2, "0");
-        const minutes = parts[1].padStart(2, "0");
-        return `${hours}:${minutes}`;
-      }
-      return timeString;
-    },
-
     // API 데이터를 컴포넌트에서 사용하는 구조로 변환
     normalizeTickets(list) {
       return list.map((t) => {
-        const { date, dayOfWeek } = this.extractDateAndDayOfWeek(t.departureDate);
+        const { date, dayOfWeek } = extractDateAndDayOfWeek(t.departureDate);
 
         return {
           id: t.airlineId,
@@ -70,14 +41,14 @@ export const useAirTicketStore = defineStore("airTicket", {
           departure: {
             city: t.departure.city,
             code: t.departure.airportCode,
-            time: this.normalizeTime(t.departure.time),
+            time: normalizeTime(t.departure.time),
           },
 
           // 도착 정보
           arrival: {
             city: t.arrival.city,
             code: t.arrival.airportCode,
-            time: this.normalizeTime(t.arrival.time),
+            time: normalizeTime(t.arrival.time),
           },
 
           // 항공편 정보
